bot/requirements.txt
aiogram~=3.20.0.post0
pybit~=5.11.0
python-dotenv==0.19.0
requests==2.26.0
notion-client==2.4.0
